python-service/requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
pydantic==2.7.2
python-dotenv==1.0.1